ROX-34178: refactor file activity + splunk TA tests
